Added new dataset for use in analysis
New dataset includes permutations for knots with small signatures (between
-5 and 5) as features along with their signatures as labels. The dataset
is saved as a .pkl file and is a 4-tuple where the train features as the
0th entry, test features as the 1st, train labels as the 2nd, and test
labels as the 3rd. This dataset is safe and can be loaded with pickle.

Also includes the jupyter notebook used to create this dataset. Please
note that the words "labels" is spelled as "lables." This is wrong,
please fix.