Modified small_signature_dataset
Instead of small_signature_dataset being a tuple of lists, it is
now a numpy array, where the features are matrices of floats and
the labels are vectors of integers. It also modifies the
data_processor appropriately and adds a helper file called
petal_assistant to assist in the processing.